requirements.txt
adlfs==2024.7.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.2
aiohttp==3.10.8
aiosignal==1.3.1
alembic==1.13.3
aniso8601==9.0.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==2.4.1
asyncer==0.0.7
attrs==23.2.0
azure-common==1.1.28
azure-core==1.31.0
azure-cosmos==4.7.0
azure-datalake-store==0.0.53
azure-identity==1.19.0
azure-search-documents==11.5.1
azure-storage-blob==12.23.1
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
bert-score==0.3.13
bidict==0.23.1
blinker==1.8.2
bokeh==3.6.0
boto3==1.35.29
botocore==1.35.29
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
chevron==0.14.0
chroma-hnswlib==0.7.6
click==8.1.7
cloudpickle==2.2.1
cohere==5.11.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
databricks-sdk==0.33.0
dataclasses-json==0.6.7
datasets==3.0.1
debugpy==1.8.6
decorator==5.1.1
deepdiff==8.0.1
deeplake==3.9.26
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
docker==7.1.0
docopt==0.6.2
docx2txt==0.8
durationpy==0.9
effdet==0.4.1
emoji==2.13.2
et-xmlfile==1.1.0
evaluate==0.4.3
execnet==2.1.1
executing==2.1.0
faiss-cpu==1.8.0
fastapi==0.112.4
fastavro==1.9.7
filelock==3.16.1
filetype==1.2.0
FlashRank==0.2.9
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.6.1
giskard==2.15.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.20.0
google-auth==2.35.0
google-cloud-vision==3.7.4
google-pasta==0.2.0
googleapis-common-protos==1.65.0
graphene==3.3
graphql-core==3.2.4
graphql-relay==3.2.0
greenlet==3.1.1
griffe==0.48.0
gritql==0.1.5
groq==0.11.0
grpcio==1.66.2
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.2
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.25.1
humanfriendly==10.0
humbug==0.3.2
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.4.5
iniconfig==2.0.0
iopath==0.1.10
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
kubernetes==31.0.0
langchain==0.3.2
langchain-community==0.3.1
langchain-core==0.3.9
langchain-experimental==0.3.2
langchain-groq==0.2.0
langchain-nomic==0.1.3
langchain-openai==0.2.2
langchain-text-splitters==0.3.0
langchain-unstructured==0.1.5
langdetect==1.0.9
langserve==0.3.0
langsmith==0.1.129
layoutparser==0.3.4
Lazify==0.4.0
literalai==0.0.607
llvmlite==0.43.0
loguru==0.7.2
lxml==5.3.0
lz4==4.3.3
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mixpanel==4.10.1
mlflow==2.16.2
mlflow-skinny==2.16.2
mmh3==5.0.1
mock==4.0.3
monotonic==1.6
mpmath==1.3.0
msal==1.31.0
msal-extensions==1.2.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
nomic==3.1.2
num2words==0.5.13
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
omegaconf==2.3.0
onnx==1.16.1
onnxruntime==1.19.2
openai==1.50.2
opencv-python==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orderly-set==5.2.2
orjson==3.10.7
overrides==7.7.0
packaging==23.2
pandas==2.2.3
parameterized==0.9.0
parso==0.8.4
pathos==0.3.2
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pi_heif==0.18.0
pikepdf==9.3.0
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
portalocker==2.10.1
posthog==3.7.0
pox==0.3.5
ppft==1.7.6.9
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==4.25.5
psutil==6.0.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
PyMuPDF==1.24.11
pynndescent==0.5.13
pypandoc==1.13
pyparsing==3.1.4
pypdf==5.0.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pysbd==0.3.4
pytest==8.3.3
pytest-repeat==0.9.3
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-engineio==4.9.1
python-iso639==2024.4.27
python-magic-bin==0.4.14
python-multipart==0.0.9
python-oxmsg==0.0.1
python-pptx==1.0.2
python-socketio==5.11.4
pytz==2024.2
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
ragas==0.1.19
RapidFuzz==3.10.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.8.1
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.5
sagemaker==2.232.2
sagemaker-core==1.0.10
sagemaker-mlflow==0.1.0
schema==0.7.7
scikit-learn==1.5.2
scipy==1.11.4
sentence-transformers==3.2.0
sentry-sdk==2.15.0
setuptools==75.1.0
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
smdebug-rulesconfig==1.0.1
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
sqlparse==0.5.1
sse-starlette==1.8.2
stack-data==0.6.3
starlette==0.37.2
sympy==1.13.3
syncer==2.0.3
tabulate==0.9.0
tblib==3.0.0
tenacity==8.4.2
threadpoolctl==3.5.0
tiktoken==0.7.0
timm==1.0.9
tokenizers==0.20.0
tomli==2.0.2
tomlkit==0.12.5
torch==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.1
typer==0.9.4
types-requests==2.32.0.20240914
typing==3.7.4.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.6
unstructured==0.15.13
unstructured-client==0.25.9
unstructured-inference==0.7.36
unstructured-ingest==0.0.19
unstructured.pytesseract==0.3.13
uptrace==1.27.0
urllib3==2.2.3
waitress==3.0.0
watchfiles==0.20.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.0.4
widgetsnbextension==4.0.13
win32-setctime==1.1.0
wrapt==1.16.0
wsproto==1.2.0
xlrd==2.0.1
XlsxWriter==3.2.0
xxhash==3.5.0
xyzservices==2024.9.0
yarl==1.13.1
zipp==3.20.2
zstandard==0.23.0
